requirements.txt
# Core dependencies
libvirt-python>=9.0.0
paramiko>=3.0.0
pydantic>=2.0.0
typer>=0.9.0
rich>=13.0.0

# Async support
asyncio-mqtt>=0.13.0
aiofiles>=23.0.0

# Web API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.11.0
sqlite3

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Logging and monitoring
structlog>=23.0.0
psutil>=5.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
httpx>=0.24.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.4.0
pre-commit>=3.3.0
